backend/pyproject.toml
[tool.poetry]
name = "your-virtual-env"
version = "0.1.0"
description = "Your Description"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pymongo = ">=4.9.0,<4.10.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
boto3 = "^1.35.70"
botocore = "^1.35.70"
backoff = "^2.2.1"
requests = "^2.32.3"
pyyaml = "^6.0.2"
langgraph = "^0.2.69"
langchain-core = "^0.3.33"
langchain-aws = "^0.2.11"
langchain-community = "^0.3.16"
grandalf = "^0.8"
langgraph-checkpoint-mongodb = "^0.1.0"
langchain-mongodb = "^0.4.0"
tqdm = "^4.67.1"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
